*** NON WORKING BUILD ***
This file integrates miniz in the manner in which the developers of miniz 
intended. However, there are key data structures and capabilities the library
is missing. (Specifically the gz_stream APIs used when transforming a channel.)

This branch is for posterity. Though if anyone can salvage the effort, have at it.